# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EJ Learning adaptive learning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teacher/schedule:
    get:
      operationId: getTeacherSchedule
      tags: [teacher]
      summary: A class's calendar days and subject slots, including blank and weekend rows
      parameters:
        - name: classId
          in: query
          required: true
          schema:
            type: integer
        - name: subjectId
          in: query
          required: false
          description: >
            Narrow to one subject the teacher holds in this class. Omitted
            means every subject they hold there, which for a class teacher or
            a primary-grade teacher is every subject the class runs.
          schema:
            type: integer
        - name: from
          in: query
          required: false
          description: Inclusive start date (YYYY-MM-DD). Defaults to seven days ago.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: to
          in: query
          required: false
          description: Inclusive end date (YYYY-MM-DD). Defaults to fourteen days ahead.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        "200":
          description: Scheduled days
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherSchedule"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /student/quiz-attempts:
    post:
      operationId: submitQuizAttempt
      tags: [student]
      summary: Record the signed-in student's run at a lesson's practice check
      description: >
        The server scores nothing: the questions live in the frontend for now,
        so it records what was answered and trusts the client's marking. That
        is fine for practice a student marks themselves and would not be for a
        graded assessment - the diagnostic block moves questions into the
        database and the scoring with them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "201":
          description: Attempt recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttempt"
        "400":
          description: Malformed attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: The account is not linked to a student record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /teacher/quiz-attempts:
    get:
      operationId: getTeacherQuizAttempts
      tags: [teacher]
      summary: What a class answered, newest first
      parameters:
        - name: classId
          in: query
          required: true
          schema:
            type: integer
        - name: subjectId
          in: query
          required: false
          description: >
            Narrow to one subject the teacher holds in this class. Omitted
            means every subject they hold there, which for a class teacher or
            a primary-grade teacher is every subject the class runs.
          schema:
            type: integer
        - name: from
          in: query
          required: false
          description: >
            Earliest calendar day to include, YYYY-MM-DD, read in
            Asia/Ulaanbaatar. Omitted means no lower bound.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: to
          in: query
          required: false
          description: Latest calendar day to include, inclusive.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Recorded attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherQuizAttempts"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /teacher/lessons:
    get:
      operationId: getTeacherLessons
      tags: [teacher]
      summary: Approved lessons for a class, in the order the book teaches them
      parameters:
        - name: classId
          in: query
          required: true
          schema:
            type: integer
        - name: subjectId
          in: query
          required: false
          description: >
            Narrow to one subject the teacher holds in this class. Omitted
            means every subject they hold there, which for a class teacher or
            a primary-grade teacher is every subject the class runs.
          schema:
            type: integer
      responses:
        "200":
          description: Lessons the class can be scheduled
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SchedulableLesson"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /teacher/schedule/generate:
    post:
      operationId: generateSchedule
      tags: [teacher]
      summary: Fill a term's empty school days from the book's own order
      description: >
        The sequence is already fixed by the textbook and the curriculum, so it
        is laid out rather than typed in. Days that already carry a lesson are
        left alone: regenerating must never discard a teacher's correction.
        Weekends are skipped; holidays are not modelled yet, so a teacher
        clears those days by hand.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateScheduleInput"
      responses:
        "200":
          description: What was laid out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateScheduleResult"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /teacher/schedule/day:
    put:
      operationId: setScheduleDay
      tags: [teacher]
      summary: Set, replace or clear one day's lesson
      description: >
        The teacher's correction surface. A null lessonId clears the day, which
        is how a holiday or a school event is recorded. Weekend lessons may only
        be assigned, changed, or cleared by an administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleDayInput"
      responses:
        "204":
          description: Day updated
        "400":
          description: Malformed day or a lesson that is not schedulable here
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /student/quiz/{lessonId}:
    get:
      operationId: getQuizPaper
      tags: [student]
      summary: The questions for a lesson, without the answers
      description: >
        The key is never sent. It used to live in the frontend bundle, where
        any student could read it, and it stays on the server now so that a
        score means something.
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizPaper"
        "403":
          description: A lesson this student is not assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
...
  schemas:
...
    SchedulableLesson:
      type: object
      properties:
        id:
          type: integer
        lessonCode:
          type: string
        lessonType:
          type: string
        skillName:
          type: string
        chapterTitle:
          type: ["string", "null"]
        pageFrom:
          type: ["integer", "null"]
      required: [id, lessonCode, lessonType, skillName, chapterTitle, pageFrom]
    GenerateScheduleInput:
      type: object
      properties:
        classId:
          type: integer
        subjectId:
          type: ["integer", "null"]
          description: >
            Lay out one subject's term. Null means every subject the teacher
            holds in this class.
        termId:
          description: >
            Which term to lay out. Omit it for the term today falls in, which
            is what a teacher pressing the button means.
          type: ["integer", "null"]
      required: [classId]
    GenerateScheduleResult:
      type: object
      properties:
        created:
          type: integer
        skipped:
          description: School days that already carried a lesson and were left alone.
          type: integer
        lessonsAvailable:
          type: integer
        firstDay:
          type: ["string", "null"]
        lastDay:
          type: ["string", "null"]
        notice:
          type: string
      required: [created, skipped, lessonsAvailable, firstDay, lastDay, notice]
    ScheduleDayInput:
      type: object
      properties:
        classId:
          type: integer
        subjectId:
          type: ["integer", "null"]
          description: >
            Which subject's day this is. Setting a lesson takes the subject
            from the lesson itself, so this only matters when clearing: without
            it, emptying Tuesday in the maths timetable would also empty
            Tuesday's physics. Null clears every subject the teacher holds in
            the class.
        scheduledOn:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        lessonId:
          description: null clears the day.
          type: ["integer", "null"]
        note:
          description: >
            What the teacher wants the class to know about this day - which
            pages to read, which exercises to do, what to watch out for. The
            student sees it. Leave the field out to keep whatever note is
            already there; send null or an empty string to remove it. Clearing
            the day removes the note with it.
          type: ["string", "null"]
          maxLength: 2000
      required: [classId, scheduledOn, lessonId]
    QuizOption:
      type: object
      properties:
        optionId:
          type: integer
        text:
          type: string
      required: [optionId, text]
    QuizQuestion:
      type: object
      properties:
        itemId:
          type: integer
        prompt:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizOption"
      required: [itemId, prompt, options]
...
    QuizPaper:
      type: object
      properties:
        lessonId:
          type: integer
        lessonCode:
          type: string
        skillName:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        kind:
          description: >
            Which sort of assessment this is - the check at the end of a
            lesson, a unit test, a monthly one, a diagnostic. Recorded against
            the lesson that carries the questions; it is not yet an assessment
            of its own, with an owner and a window.
          type: string
          enum: [LESSON, UNIT, MONTHLY, DIAGNOSTIC]
        takenToday:
          description: >
            Whether this student has already sat this quiz today. A quiz may be
            taken once a day, so the paper says so up front rather than letting
            a child answer everything again and be refused at the end.
          type: boolean
        previousScore:
          type: ["integer", "null"]
        previousMaxScore:
          type: ["integer", "null"]
      required:
        - lessonId
        - lessonCode
        - skillName
        - questions
        - kind
        - takenToday
        - previousScore
        - previousMaxScore
    QuizAnswer:
      type: object
      description: What the teacher sees afterwards, with the text copied in so it still reads after an edit.
      properties:
        questionId:
          type: string
        prompt:
          type: string
        chosenOptionId:
          type: string
        chosenText:
          type: string
        correct:
          type: boolean
      required: [questionId, prompt, chosenOptionId, chosenText, correct]
    QuizAttemptInput:
      type: object
      description: Only the choices. The server looks up which were right.
      properties:
        lessonId:
          type: integer
        answers:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              itemId:
                type: integer
              optionId:
                type: ["integer", "null"]
            required: [itemId, optionId]
      required: [lessonId, answers]
    QuizResult:
      type: object
      properties:
        itemId:
          type: integer
        correct:
          type: boolean
        correctOptionId:
          type: ["integer", "null"]
        explanation:
          type: ["string", "null"]
      required: [itemId, correct, correctOptionId, explanation]
    QuizAttempt:
      type: object
      properties:
        id:
          type: integer
        lessonCode:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        submittedAt:
          type: string
        results:
          description: Marking comes back with the attempt, which is the first time the key is disclosed.
          type: array
          items:
            $ref: "#/components/schemas/QuizResult"
      required: [id, lessonCode, score, maxScore, submittedAt, results]
    TeacherQuizAttemptRow:
      type: object
      properties:
        id:
          type: integer
        studentId:
          description: Needed to assign this student extra work straight from the row.
          type: integer
        studentName:
          type: string
        studentCode:
          type: string
        lessonCode:
          type: string
        skillName:
          type: string
        kind:
          description: >
            Which sort of assessment this sitting was - the check at the end of
            a lesson, a unit test, a monthly one, a diagnostic. A teacher
            reading a month of results needs the monthly test told apart from
            the daily checks.
          type: string
          enum: [LESSON, UNIT, MONTHLY, DIAGNOSTIC]
        score:
          type: integer
        maxScore:
          type: integer
        submittedAt:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
      required:
        - id
        - studentId
        - studentName
        - studentCode
        - lessonCode
        - skillName
        - kind
        - score
        - maxScore
        - submittedAt
        - answers
    TeacherQuizAttempts:
      type: object
      properties:
        classId:
          type: integer
        className:
          type: string
        from:
          description: The range actually applied, null where there was no bound.
          type: ["string", "null"]
        to:
          type: ["string", "null"]
        truncated:
          description: >
            True when the limit cut the list short, so the totals below it are
            of what came back rather than of the range asked for.
          type: boolean
        attempts:
          type: array
          items:
            $ref: "#/components/schemas/TeacherQuizAttemptRow"
      required: [classId, className, from, to, truncated, attempts]
...
    ScheduledDay:
      type: object
      properties:
        scheduledOn:
          description: Calendar date, YYYY-MM-DD.
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        isToday:
          type: boolean
        subjectId:
          description: >
            The subject this row's lesson belongs to, null on a day nothing is
            scheduled for. A day appears once per subject taught that day, so
            this is what tells two rows of the same date apart.
          type: ["integer", "null"]
        subject:
          type: ["string", "null"]
        lessonId:
          description: Null on a teaching day nothing is scheduled for yet.
          type: ["integer", "null"]
        lessonCode:
          type: ["string", "null"]
        lessonType:
          type: ["string", "null"]
        skillName:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
      required:
        [scheduledOn, isToday, subjectId, subject, lessonId, lessonCode,
         lessonType, skillName, note]
    TeacherSchedule:
      type: object
      properties:
        classId:
          type: integer
        className:
          type: string
        gradeLevel:
          type: integer
        stage:
          type: string
          enum: [PRIMARY, SECONDARY]
          description: >
            Derived from the class's grade, not stored on the teacher: the two
            workflows differ per class, and one teacher may hold both.
        days:
          type: array
          items:
            $ref: "#/components/schemas/ScheduledDay"
      required: [classId, className, gradeLevel, stage, days]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
...